Add SPICE specific changes to new git branch
