+++ cpm1/ML_models/mk1/assimilate.py
@@ -2,6 +2,8 @@
 
 # Find a point in latent space that maximises the fit to some given input fields,
 #  and plot the fitted state.
+    #
+    # assimilate.py --epoch 500 --year 1990 --day 786
 
 import os
 import sys
@@ -9,6 +11,7 @@ import numpy as np
 import tensorflow as tf
 import tensorflow_probability as tfp
 
+sys.path.append('/home/h03/hadsx/extremes/ML/pb1/DCVAE_Climate_sjb1/cpm1')
 from ML_models.mk1.makeDataset import getDataset
 from ML_models.mk1.autoencoderModel import getModel
 
@@ -20,11 +23,9 @@ import argparse
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--epoch", help="Epoch", type=int, required=False, default=500)
+parser.add_argument("--year", help="Test year", type=int, required=False, default=1990)
 parser.add_argument(
-    "--year", help="Year to fit to", type=int, required=False, default=None
-)
-parser.add_argument(
-    "--month", help="Month to fit to", type=int, required=False, default=None
+    "--day", help="Test day", type=int, required=False, default=786
 )
 for field in specification["outputNames"]:
     parser.add_argument(
@@ -33,13 +34,13 @@ for field in specification["outputNames"]:
         default=False,
         action="store_true",
     )
-for field in specification["outputNames"]:
-    parser.add_argument(
-        "--%s_mask" % field,
-        help="Mask for fit to %s?" % field,
-        type=str,
-        default=None,
-    )
+# for field in specification["outputNames"]:
+#     parser.add_argument(
+#         "--%s_mask" % field,
+#         help="Mask for fit to %s?" % field,
+#         type=str,
+#         default=False,  # None,
+#     )
 parser.add_argument(
     "--iter",
     help="No. of iterations",
@@ -57,11 +58,11 @@ args = parser.parse_args()
 args_dict = vars(args)
 # Load the masks, if specified
 fitted = {}
-masked = {}
+# masked = {}
 for field in specification["outputNames"]:
-    masked[field] = None
-    if args_dict["%s_mask" % field] is not None:
-        masked[field] = np.load(args_dict["%s_mask" % field])
+    # masked[field] = None
+    # if args_dict["%s_mask" % field] is not None:
+    #     masked[field] = np.load(args_dict["%s_mask" % field])
     fitted[field] = args_dict[field]
 
 purpose = "Test"
@@ -78,16 +79,20 @@ year = None
 month = None
 for batch in dataset:
     dateStr = tf.strings.split(batch[0][0][0], sep="/")[-1].numpy()
-    year = int(dateStr[:4])
-    month = int(dateStr[5:7])
-    if (args.month is None or month == args.month) and (
-        args.year is None or year == args.year
-    ):
+    # print(dateStr)
+    # year = int(dateStr[:4])
+    # month = int(dateStr[5:7])
+    year = int(dateStr[10:14]) # int(fN[:4])
+    # month = int(fN[5:7])
+    idot    = str(dateStr).find('.')
+    day     = int(dateStr[15:(idot-2)]) # int(fN[5:7])
+    if (day == args.day) and (year == args.year):
         input = batch
         break
+    # input = batch
 
 if input is None:
-    raise Exception("Month %04d-%02d not in %s dataset" % (year, month, purpose))
+    raise Exception("%04d-%02d not in %s dataset" % (year, day, purpose))
 
 autoencoder = getModel(specification, args.epoch)
 
@@ -107,21 +112,22 @@ def decodeFit():
     for field in specification["outputNames"]:
         if fitted[field]:
             field_idx = specification["outputNames"].index(field)
-            mask = masked[field]
-            if mask is not None:
-                mask = tf.constant(mask, dtype=tf.float32)
-                result = result + tf.reduce_mean(
-                    tf.keras.metrics.mean_squared_error(
-                        tf.boolean_mask(generated[0, :, :, field_idx], mask),
-                        tf.boolean_mask(target[:, :, field_idx], mask),
-                    )
+            # mask = masked[field]
+            # mask = None
+            # if mask is not None:
+            #     mask = tf.constant(mask, dtype=tf.float32)
+            #     result = result + tf.reduce_mean(
+            #         tf.keras.metrics.mean_squared_error(
+            #             tf.boolean_mask(generated[0, :, :, field_idx], mask),
+            #             tf.boolean_mask(target[:, :, field_idx], mask),
+            #         )
+            #     )
+            # else:
+            result = result + tf.reduce_mean(
+                tf.keras.metrics.mean_squared_error(
+                    generated[:, :, :, field_idx], target[:, :, field_idx]
                 )
-            else:
-                result = result + tf.reduce_mean(
-                    tf.keras.metrics.mean_squared_error(
-                        generated[:, :, :, field_idx], target[:, :, field_idx]
-                    )
-                )
+            )
     return result
 
 
@@ -138,4 +144,6 @@ if any(fitted.values()):
 generated = autoencoder.generate(latent, training=False)
 
 # Make the plot - same as for validation script
-plotValidationField(specification, input, generated, year, month, "assimilated.webp")
+title0="%s e%d %04d-%3d" % (specification["modelName"], args.epoch, year, day)
+savefile0="ass_%s_e%d_%04d-%3d.webp" % (specification["modelName"], args.epoch, year, day)
+plotValidationField(specification, input, generated, year, day, savefile0, title0)
