+++ cpm1/make_raw_tensors/CPM5/make_all_tensors.py
@@ -11,7 +11,7 @@ sDir = os.path.dirname(os.path.realpath(__file__))
 parser = argparse.ArgumentParser()
 parser.add_argument("--member", help="Member", type=int, required=True)
 parser.add_argument("--variable", help="Variable name", type=str, required=True)
-# parser.add_argument("--year", help="Year", type=int, required=True)
+parser.add_argument("--year", help="Year", type=int, required=True)
 # parser.add_argument("--day", help="Integer day", type=int, required=True)
 args = parser.parse_args()
 # print(args)
@@ -35,21 +35,23 @@ def is_done(member, variable, year, day):
 
 
 count = 0
-for year in range(1980, 1981):
-    # for day in range(0, 1000):
-    for day in range(0, 900):
+# for year in range(1980, 1981):
+year = args.year
+for day in range(0, 900):
+# for day in range(0, 2):
+    # print(year, day)
 
-    # for day in range(0, 11):
-        if is_done(args.member, args.variable, year, day):
-            continue
-        cmd = "%s/make_training_tensor.py --member=%d --variable=%s --year=%04d --day=%d" % (
-            sDir,
-            args.member,
-            args.variable,
-            year,
-            day,
-        )
-        print(cmd)
+# for day in range(0, 11):
+    # if is_done(args.member, args.variable, year, day):
+    #     continue
+    cmd = "%s/make_training_tensor.py --member=%d --variable=%s --year=%04d --day=%d" % (
+        sDir,
+        args.member,
+        args.variable,
+        year,
+        day,
+    )
+    print(cmd)
 
 # time.sleep(20)
