+++ cpm1/make_raw_tensors/make_all_raw_tensors.sh
@@ -5,9 +5,13 @@
 
 ### make_all_raw_tensors.sh | ~hadsx/bin/spice_parallel --time=5
 
-# (cd CPM5 && ./make_all_tensors.py --member=1 --variable=tas)
-(cd CPM5 && ./make_all_tensors.py --member=1 --variable=psl)
-sleep 60
-(cd CPM5 && ./make_all_tensors.py --member=1 --variable=uas)
-sleep 60
-(cd CPM5 && ./make_all_tensors.py --member=1 --variable=vas)
+### cd CPM5
+### ./make_all_tensors.py --member=1 --variable=uas --year=1990 | ~hadsx/bin/spice_parallel --time=5
+
+(cd CPM5 && ./make_all_tensors.py --member=1 --variable=tas --year=1990)
+# sleep 60
+(cd CPM5 && ./make_all_tensors.py --member=1 --variable=psl --year=1990)
+# sleep 60
+(cd CPM5 && ./make_all_tensors.py --member=1 --variable=uas --year=1990)
+# sleep 60
+(cd CPM5 && ./make_all_tensors.py --member=1 --variable=vas --year=1990)
