+++ cpm1/normalize/normalise_all.sh
@@ -3,7 +3,7 @@
 # Make normalization constants for all the datasets
 # Requires pre-made raw tensors
 
-### normalise_all.sh | ~hadsx/bin/spice_parallel --time=5
+### normalise_all.sh | ~hadsx/bin/spice_parallel --time=10
 
 
 (cd CPM5 && ./make_all_fits.py)
